add naive image downloading for saved pages
I'm considering this naive because the images will be inflated and put
into memory—let alone downloading them in the background.  We need to at
least skip the memory cache for these and allow them to be fetched at
low priority in the background.

also made `allImageURLs` a set to enforce unique-ness when adding lead &
thumbnail images.