Show the previous VC when launching the app "cold"
- Track the "open article"
This is done using NSUserDefaults - and setting it from the container VC

- Check for the last open article on launch (instead of the most recent article)
This checks the value in NSUserDefaults.

- Make sure we don't "resume" the app before the launch sequence completes (and the view hierarchy is setup)
The notification for resuming the app can come before the app launch sequence occurs (migration, splash screen, on boarding… etc…)
This is now accounted for by checking "launchCompleted" before executing the reumeApp method
This was necessary as we might try to push the last open article before the UI is ready.