fix gallery crash by removing image import side effect from webview
Image downloads from the webview were causing image objects to be added
to the "current article."  This is wrong on many levels (i.e. thread
safety, side effects, global state), and needed to be removed now that
we have proper image caching anyway.  In this case, it caused the
article's image list to change in the time between when the header
gallery was created and the fullscreen gallery was created.