Update languages VC to show/hide preferred languages.
Also removed the logic for adding a preferred language from the VC and deferred that to the delegate.
This was necessary because not all language selections affect the preferred language list in the same way.

Updated all the delegates to account for new behavior.
Refactor delegate method signature to b more idiomatic.