Move VoiceOver cursor to place in article on TOC item selection
This tries to move the VoiceOver cursor to the correct place in article
when the user selects an item in the Table of Contents.

However, the handling of VoiceOver cursor receival seems to be buggy in
UIWebView and therefore this commit is only a partial solution.  It
just implements a way that resolves the situation in the majority of
cases.