Fix edit preview not using preview.html and thus not using preview.css.
Basically we want to use preview.html when previewing an edit, but
instead it would just keep using index.html because we had added an
optimization which didn't take into account that the method had a
parameter which changes what html asset to use.

That meant when previewing edits we weren't using preview.css either, and this caused the weird EDIT link to be visible (among other things).