Update nearby section insertion to throw away cached data
CLLocationmanager can return a cached location - we want the new one. So we check the time stamp to make sure its within the last 5 minutes before we use it.
Additionally, we stop the location manager because it doesnt need to run (we will restart it the next time we update)