always call "willDisplaySection" whenever a cell will be displayed
in the end, we really don't need to only call "start" once, as long as
the controllers can make it idempotent (which is easy enough in nearby's
case).  the table view just gets into too many situations where the
"willDisplay" is called after more than just the first cell is on
screen, so we can't try to be clever about it.  "didEndDisplaying" is a
simpler case, since it's either 1 or 0 cells visible, which is much more
reliable