Adding image caching functionality to WMFURLCache
- Add method to each all images for an article into the permanent cache (SDWebImage)
- Add method to retrieve a cached image for a given url

Doing this allows us to use the built in url cache,
BUT when we need to save an article we can query the url cache for any article images that we need to permanently cache.

The impetus for this was that we were decoding large gifs and we were paying triple penalties for the image cost.
Moreover the memory spiked (as opposed to the behavior now, where it grows gradually as each frame is decoded)