Mark a title as "non-standard"
This is used for identifying titles that should not be handled like normal titles.
They should not be added to history or be savable. They should not be stored as the last open article.
This is so we can use the article VC to display non-wikipedia articles like the FAQ

Several classes were updated accordingly