Fix the captcha presentation code
We need to differentiate between the first attempt and subsequent attempts to solve a captcha.

In reality, the user never enters text to solve the captcha on the first attempt. We immediately try to create an account with the supplied credentials.
This is because we want to get any errors back on the username (is it a duplicate, is it valid) before letting them solve the captcha.

So on the first attempt, we try and fail, once we get the captcha error, we present the captcha to the user
On subsequent attempts, if we fail it is because the user actually messed up the captcha, and in this case we need to request a new captcha for the user.