New vc & storyboard for 2 factor login. New currently logged in user fetcher.
Because 2FA accounts *always* require 2FA token on login attempts
we can't just always try to log in with saved credentials on app start.
( We'd have to bug the user for 2FA token each time. )
Instead ask the API via "action=query&meta=userinfo" to see if a user
is logged in.