Fix bug preventing "password fields do not match" message from ever appearing.
The problem was caused by the password match check happening as part
of the done button "enabled" logic. Imagine the user had typed 2 passwords
each having, say, 8 characters. Now imagine the 8 characters in the 2
password fields didn't match. They would see 8 "dots" in each of the
passwords fields, but the done button would not be enabled (because
we were checking for the match in the done button enabled logic). So
with the done button disabled, in this scenario they would never see the
"passwords didn't match" message. The fix in this commit allows the done
button to be enabled if the passwords don't match, which allows the user
to tap the "done" button in such cases, which allows the "guard" in the
save method to present a "passwords don't match" message when
needed.