Fix for T153909: Header image appears too large on iPad
We determined that the header image is already appearing at the correct height.

We also decided that the image should be top-aligned, which is non-trivial unfortunately. Some quick research lead me to a small project called UIImageViewAligned [1]. I just copied the header and implementation files into the WMF framework and it seems to be working well.

 [1] https://github.com/reydanro/UIImageViewAligned