Rebuild captcha storyboard to use stackview so we can easily hide things.
Stub out outlets for title and sub-title which will be moved into the captcha
vc.

Neat thing I didn't know about stack views:

If a view is placed in a stackview cell and constrained by that cell, you
can set the view to be hidden and the stackview will remove it from the
view hierarchy *yet retain it*. When this happens the stackview handles
adjusting constraints so the cell collapses. To re-show the view you just
set hidden to false and the stackview re-inserts the cell into the view
hierarchy and adjusts the constraints automatically:
http://www.iosinsight.com/uistackview-for-showing-and-hiding-user-interface-controls/