Hook up keyboard "Done" key.
This is a little odd because iOS doesn't provide an easy way to keep the keyboard return key disabled until, in this case, 16 characters have been entered.

Joe reminded me that the backup code length may change in the future, which Gergo had mentioned as well.
So for now let's not enforce a min length on the backup code.