Make img widening transform properly detect images which should not be widened.
Was formerly dependent on whether an image overflow container div was added,
but the proper way to see if an image shouldn't be resized is by checking for the
"noresize" class.

Example of such an image is the one on "enwiki > Counties of
England" which has county links overlaying an image of England. That image has
a "noresize" class on its div because the overlay links are positioned absolutely,
so you don't want to resize the image or the links would be in the wrong place.