first pass at wiring up text size
html tags are a bit easier to handle separately from wikitext markup items.
the approach introduced provides a method for mapping the current line's tokens into overall ranges for the tags on that line. these are vastly simpler to work with than the large number of tokens which could comprise a given tag's contents. for example, it's much easier to compare these overall ranges with the selection range