Present language variant alerts when needed - Add WMFAppViewController+Extensions.swift file to project - Add user defaults key to track latest library version where migration occurred - Add method to present all required language variant alerts - Add presentVariantAlert() method to present single alert, with handlers to potentially   present addtional alerts, dismiss, or navigate to langague preferences
Note that I added the presentation methods here instead of Panels.swift in part because they
require an extension to WMFAppViewController and in part because of the very large length of
Panels.swift. This seemed to keep all of the presentation logic in one place.

Note that I added the WMFAppViewController+Extensions.swift file for two reasons. The first is
to prevent needing to adapt education panels for Obj-C usage. The second is to provide a place
for new WMFAppViewController code to go, and be written in Swift.