UserHstoryFunnel to send data to MEP
This commit fully transitions the UserHistoryFunnel to MEP
To do that, it removes compliance to legacy protocols, and removes the use of the legacy class EventLoggingService. Other than changing how the data is sent and how the funnel is initialized, it also changes the way we save the funnel snapshot. Since the new Event type is not compliant with NSCoding - the way it was done before, but is compliant to codable, it allowed us to move from saving this snapshot (a copy of the event used for comparison purposes, so we don't send data that hasn't changed) in userDefaults to using our caching service that complies with the codable protocol.