Set shouldAttemptToShowArticleAsLivingDoc prop to false
This variable was causing the explore feed to crash when hitting on some cards. It happens after the removal of calls to WMFLoggingService on WMFAppViewController setupControllers method. The cause and effect relantionship was not clear, so I'm changing this variable to always return false, since we are removing the Article as Living doc code in the near future.