Fix search source bug on ios search funnel.
The previous approach of casting the view controller to the new protocol was always failing, so now we use
 the count of view controllers on the stack.
For the explore feed, the search will always have a count of two view controllers. For the article, from navigating on the app and deep linking from outside the app, the count will be at least three.