Migrate persisted preferred languages to new codes
This commit adds mapping to all the language variants being updated to the new codes. If the current user settings has any of them stored, it will be updated correctly to the new code.
new variants added in the future with the correct code probably won't need to be migrated.